+++ 00 - Tony/20FreemindCommon/62_fxCharting/StockSharp/Xaml/Charting/00 - Freemind/FreemindCandlestickRenderableSeries_Reduced.cs
@@ -79,7 +79,7 @@ namespace fx.Charting
                 var tmpBar        = ( SBar ) metadata;
                 CheckBarsRepo( ref tmpBar );
                 ref SBar realBar  = ref _barList[ tmpBar.Index ];
-                realBarIndex      = ( int ) realBar.Index;
+                realBarIndex      = realBar.Index;
 
                 if ( realBar.IsSelected )
                 {
@@ -307,7 +307,7 @@ namespace fx.Charting
                 var tmpBar        = ( SBar ) metadata;
                 CheckBarsRepo( ref tmpBar );
                 ref SBar realBar  = ref _barList[ tmpBar.Index ];
-                var realBarIndex  = ( int ) realBar.Index;
+                var realBarIndex  = realBar.Index;
 
                 if ( realBar.IsSelected )
                 {
@@ -364,7 +364,7 @@ namespace fx.Charting
                 var tmpBar        = ( SBar ) metadata;
                 CheckBarsRepo( ref tmpBar );
                 ref SBar realBar  = ref _barList[ tmpBar.Index ];
-                realBarIndex      = ( int ) realBar.Index;
+                realBarIndex      = realBar.Index;
 
                 if ( realBar.IsSelected )
                 {
